+++ backend/lib/actions/teacher.ts
@@ -6,119 +6,159 @@ import { statusCodes } from "../types/statusCodes";
 const prisma = new PrismaClient();
 
 function convertTableToString(timetable: string[][]): string {
-  return timetable.map(row => row.join(",")).join(";");
+  return timetable.map((row) => row.join(",")).join(";");
 }
 
-
-export async function createTeachers(JWTtoken: string, name: string, initials: string | null = null, email: string | null = null, department: string | null = null, alternateDepartments: string | null = null, timetable: string[][] | null = null, labtable: string[][] | null = null): Promise<{ status: number, teacher: Teacher | null }> {
+export async function createTeachers(
+  JWTtoken: string,
+  name: string,
+  initials: string | null = null,
+  email: string | null = null,
+  department: string | null = null,
+  alternateDepartments: string | null = null,
+  timetable: string[][] | null = null,
+  labtable: string[][] | null = null
+): Promise<{ status: number; teacher: Teacher | null }> {
   try {
-    const { status, user } = await auth.getPosition(JWTtoken)
+    const { status, user } = await auth.getPosition(JWTtoken);
+
+    if (user?.orgId == null) {
+      return {
+        status: statusCodes.BAD_REQUEST,
+        teacher: null,
+      };
+    }
+
     if (status == statusCodes.OK && user) {
       //if user isnt a viewer
-      if (user.role != 'viewer') {
+      if (user.role != "viewer") {
         //check if teacher with same name dep and org exist
         const teachers = await prisma.teacher.findFirst({
           where: {
             name: name,
             department: department ? department : user.department,
-            organisation: user.organisation
-          }
-        })
+            orgId: user.orgId,
+          },
+        });
         //if even a single teacher exists
         if (teachers) {
           return {
             status: statusCodes.BAD_REQUEST,
-            teacher: null
-          }
+            teacher: null,
+          };
         }
         //else
         const teacher: Teacher = {
           name: name,
           initials: initials,
           email: email,
-          department: department ? department : user.department ? user.department : "no department",
+          department: department
+            ? department
+            : user.department
+            ? user.department
+            : "no department",
           alternateDepartments: alternateDepartments,
-          timetable: timetable ? convertTableToString(timetable) : "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
-          labtable: labtable ? convertTableToString(labtable) : "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
-          organisation: user.organisation
-        }
+          timetable: timetable
+            ? convertTableToString(timetable)
+            : "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
+          labtable: labtable
+            ? convertTableToString(labtable)
+            : "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
+          orgId: user.orgId,
+        };
         await prisma.teacher.create({
-          data: teacher
-        })
+          data: teacher,
+        });
         return {
           status: statusCodes.CREATED,
-          teacher: teacher
-        }
+          teacher: teacher,
+        };
       }
       //if user is a viewer code will reach here
       return {
         status: statusCodes.FORBIDDEN,
-        teacher: null
-      }
+        teacher: null,
+      };
     }
     //if status not ok
     return {
       status: status,
-      teacher: null
-    }
-  }
-  catch {
+      teacher: null,
+    };
+  } catch {
     return {
       status: statusCodes.INTERNAL_SERVER_ERROR,
-      teacher: null
-    }
+      teacher: null,
+    };
   }
 }
-export async function updateTeachers(JWTtoken: string, originalName: string, originalDepartment: string|null=null, teacher: Teacher): Promise<{ status: number, teacher: Teacher | null }> {
+export async function updateTeachers(
+  JWTtoken: string,
+  originalName: string,
+  originalDepartment: string | null = null,
+  teacher: Teacher
+): Promise<{ status: number; teacher: Teacher | null }> {
   try {
     const { status, user } = await auth.getPosition(JWTtoken);
+
+    if (user?.orgId == null) {
+      return {
+        status: statusCodes.BAD_REQUEST,
+        teacher: null,
+      };
+    }
+
     if (status == statusCodes.OK && user) {
-      if (user.role != 'viewer') {
+      if (user.role != "viewer") {
         const teacherExists = await prisma.teacher.findFirst({
           where: {
             name: originalName,
-            department: user.role=='admin'?originalDepartment:user.department,
-            organisation: user.organisation
-          }
+            department:
+              user.role == "admin" ? originalDepartment : user.department,
+            orgId: user.orgId,
+          },
         });
         if (!teacherExists) {
           return {
             status: statusCodes.BAD_REQUEST,
-            teacher: null
+            teacher: null,
           };
         }
         const updatedTeacher = await prisma.teacher.update({
           where: {
-            id: teacherExists.id
+            id: teacherExists.id,
           },
           data: {
             name: teacher.name,
             initials: teacher.initials,
             email: teacher.email,
-            department:user.role=='admin' && teacher.department?teacher.department:user.department ,
+            department:
+              user.role == "admin" && teacher.department
+                ? teacher.department
+                : user.department,
             alternateDepartments: teacher.alternateDepartments,
             timetable: teacher.timetable,
             labtable: teacher.labtable,
-          }
+          },
         });
         return {
           status: statusCodes.OK,
-          teacher: updatedTeacher
+          teacher: updatedTeacher,
         };
       }
       return {
         status: statusCodes.FORBIDDEN,
-        teacher: null
+        teacher: null,
       };
     }
     return {
       status: status,
-      teacher: null
+      teacher: null,
     };
   } catch {
     return {
       status: statusCodes.INTERNAL_SERVER_ERROR,
-      teacher: null
+      teacher: null,
     };
   }
 }
@@ -132,9 +172,16 @@ export async function createManyTeachers(
 ): Promise<{ status: number; teachers: Teacher[] | null }> {
   try {
     const { status, user } = await auth.getPosition(JWTtoken);
-    if (status == statusCodes.OK && user && user.role != "viewer") {
 
-      const teachers: Teacher[] = []
+    if (user?.orgId == null) {
+      return {
+        status: statusCodes.BAD_REQUEST,
+        teachers: null,
+      };
+    }
+
+    if (status == statusCodes.OK && user && user.role != "viewer") {
+      const teachers: Teacher[] = [];
 
       for (let i = 0; i < name.length; i++) {
         teachers.push({
@@ -143,17 +190,19 @@ export async function createManyTeachers(
           email: email ? email[i] : null,
           department: department ? department : user.department,
           alternateDepartments: null,
-          timetable: "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
-          labtable: "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
-          organisation: user.organisation
-        })
+          timetable:
+            "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
+          labtable:
+            "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
+          orgId: user.orgId,
+        });
       }
 
       const duplicateChecks = await Promise.all(
         teachers.map((teacher) =>
           prisma.teacher.findFirst({
             where: {
-              organisation: teacher.organisation,
+              orgId: teacher.orgId,
               department: teacher.department,
               name: teacher.name,
             },
@@ -173,51 +222,61 @@ export async function createManyTeachers(
         data: teachers,
       });
 
-            return {
-                status: statusCodes.CREATED,
-                teachers: teachers,
-            };
-        }
-        return {
-            status: statusCodes.FORBIDDEN,
-            teachers: null,
-        };
-    } catch {
-        return {
-            status: statusCodes.INTERNAL_SERVER_ERROR,
-            teachers: null,
-        };
+      return {
+        status: statusCodes.CREATED,
+        teachers: teachers,
+      };
     }
+    return {
+      status: statusCodes.FORBIDDEN,
+      teachers: null,
+    };
+  } catch {
+    return {
+      status: statusCodes.INTERNAL_SERVER_ERROR,
+      teachers: null,
+    };
+  }
 }
-//to display teachers list 
-export async function getTeachers(JWTtoken: string): Promise<{ status: number, teachers: Teacher[] | null }> {
+//to display teachers list
+export async function getTeachers(
+  JWTtoken: string
+): Promise<{ status: number; teachers: Teacher[] | null }> {
   try {
-    const { status, user } = await auth.getPosition(JWTtoken)
+    const { status, user } = await auth.getPosition(JWTtoken);
+
+    if (user?.orgId == null) {
+      return {
+        status: statusCodes.BAD_REQUEST,
+        teachers: null,
+      };
+    }
 
     //if verification of roles is ok
     if (status == statusCodes.OK && user) {
       let teachers: Teacher[];
 
       //if role isnt admin, return teachers from same department
-      if (user.role != 'admin') {
-        teachers = await prisma.teacher.findMany({
-          where: {
-            organisation: user.organisation,
-            department: user.department
-          },
-          select: {
-            name: true,
-            department: true,
-            initials: true,
-            email: true,
-          }
-        })
+      if (user.role != "admin") {
+        teachers = await prisma.teacher
+          .findMany({
+            where: {
+              orgId: user.orgId,
+              department: user.department,
+            },
+            select: {
+              name: true,
+              department: true,
+              initials: true,
+              email: true,
+              orgId: true,
+            },
+          })
           //convert the returned object into Teacher[] type
           .then((teachers) =>
             teachers.map((teacher) => ({
               ...teacher,
               alternateDepartments: null,
-              organisation: user.organisation,
               timetable: null, // Default value, since it's not queried
               labtable: null, // Default value, since it's not queried
             }))
@@ -225,23 +284,24 @@ export async function getTeachers(JWTtoken: string): Promise<{ status: number, t
       }
       //if role is admin, return teachers from all departments
       else {
-        teachers = await prisma.teacher.findMany({
-          where: {
-            organisation: user.organisation,
-          },
-          select: {
-            name: true,
-            department: true,
-            initials: true,
-            email: true
-          }
-        })
+        teachers = await prisma.teacher
+          .findMany({
+            where: {
+              orgId: user.orgId,
+            },
+            select: {
+              name: true,
+              department: true,
+              initials: true,
+              email: true,
+              orgId: true,
+            },
+          })
           //convert the returned object into Teacher[] type
           .then((teachers) =>
             teachers.map((teacher) => ({
               ...teacher,
               alternateDepartments: null,
-              organisation: user.organisation,
               timetable: null, // Default value, since it's not queried
               labtable: null, // Default value, since it's not queried
             }))
@@ -249,22 +309,19 @@ export async function getTeachers(JWTtoken: string): Promise<{ status: number, t
       }
       return {
         status: statusCodes.OK,
-        teachers: teachers
-      }
-
-    }
-    else {
+        teachers: teachers,
+      };
+    } else {
       return {
         status: status,
-        teachers: null
-      }
+        teachers: null,
+      };
     }
-  }
-  catch {
+  } catch {
     return {
       status: statusCodes.INTERNAL_SERVER_ERROR,
-      teachers: null
-    }
+      teachers: null,
+    };
   }
 }
 
@@ -276,17 +333,30 @@ export async function peekTeacher(
   try {
     //get position of user
     const { status, user } = await auth.getPosition(token);
+
+    if (user?.orgId == null) {
+      return {
+        status: statusCodes.BAD_REQUEST,
+        teacher: null,
+      };
+    }
+
     //if verification of rules is okay, perform the following
     if (status == statusCodes.OK && user) {
       const teacher = await prisma.teacher.findFirst({
         where: {
           name: name,
-          department: user.role == 'admin' ? department ? department : user.department : user.department,
-          organisation: user.organisation,
+          department:
+            user.role == "admin"
+              ? department
+                ? department
+                : user.department
+              : user.department,
+          orgId: user.orgId,
         },
         select: {
           name: true,
-          organisation: true,
+          orgId: true,
           department: true,
           alternateDepartments: true,
           initials: true,
@@ -300,13 +370,12 @@ export async function peekTeacher(
         teacher: teacher,
       };
     }
-    //else 
+    //else
     return {
       status: status,
-      teacher: null
-    }
-  }
-  catch {
+      teacher: null,
+    };
+  } catch {
     //internal error
     return {
       status: statusCodes.INTERNAL_SERVER_ERROR,
@@ -315,37 +384,47 @@ export async function peekTeacher(
   }
 }
 
-export async function deleteTeachers(JWTtoken: string, teachers: Teacher[]): Promise<{ status: number }> {
-  const { status, user } = await auth.getPosition(JWTtoken)
+export async function deleteTeachers(
+  JWTtoken: string,
+  teachers: Teacher[]
+): Promise<{ status: number }> {
+  const { status, user } = await auth.getPosition(JWTtoken);
+
+  if (user?.orgId == null) {
+    return {
+      status: statusCodes.BAD_REQUEST,
+    };
+  }
+
   try {
     if (status == 200 && user) {
-      if (user.role != 'viewer') {
+      if (user.role != "viewer") {
         await prisma.teacher.deleteMany({
           where: {
-            OR: teachers.map(teacher => ({
+            OR: teachers.map((teacher) => ({
               name: teacher.name,
-              organisation: user.organisation,
-              department: user.role == 'admin' ? teacher.department : user.department
-            }))
-          }
-        })
+              orgId: user.orgId as number,
+              department:
+                user.role == "admin" ? teacher.department : user.department,
+            })),
+          },
+        });
         return {
-          status: statusCodes.OK
-        }
+          status: statusCodes.OK,
+        };
       }
       //else
       return {
-        status: statusCodes.FORBIDDEN
-      }
+        status: statusCodes.FORBIDDEN,
+      };
     }
     //else
     return {
-      status: status
-    }
-  }
-  catch {
+      status: status,
+    };
+  } catch {
     return {
-      status: statusCodes.INTERNAL_SERVER_ERROR
-    }
+      status: statusCodes.INTERNAL_SERVER_ERROR,
+    };
   }
 }
